# internal/api/docs/swagger.yaml
definitions:
...
  entity.Sale:
    properties:
      client_id:
        type: string
      client_name:
        type: string
      client_phone:
        type: string
      payment_method:
        type: string
      sold_products:
        items:
          $ref: '#/definitions/entity.SalesItem'
        type: array
    type: object
  entity.SaleUpdate:
    properties:
      client_id:
        type: string
      payment_method:
        type: string
    type: object
  entity.SalesItem:
    properties:
      id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      sale_id:
        type: string
      sale_price:
        type: number
      total_price:
        type: number
    type: object
...
  products.Error:
    properties:
      error:
        type: string
    type: object
...
  products.Message:
    properties:
      message:
        type: string
    type: object
...
  products.SaleList:
    properties:
      sales:
        items:
          $ref: '#/definitions/products.SaleResponse'
        type: array
    type: object
  products.SaleResponse:
    properties:
      branch_id:
        description: Added branch_id
        type: string
      client_id:
        type: string
      client_name:
        type: string
      client_phone_number:
        type: string
      company_id:
        description: Company ID added
        type: string
      created_at:
        type: string
      id:
        type: string
      payment_method:
        type: string
      sold_by:
        type: string
      sold_by_name:
        type: string
      sold_products:
        items:
          $ref: '#/definitions/products.SalesItem'
        type: array
      total_sale_price:
        description: Changed to double
        type: number
    type: object
  products.SalesItem:
    properties:
      id:
        type: string
      product_id:
        type: string
      product_image:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      sale_id:
        type: string
      sale_price:
        description: Changed to double
        type: number
      total_price:
        description: Changed to double
        type: number
    type: object
...
info:
  contact: {}
  description: This is a sample server
  title: API Gateway
  version: "1.0"
paths:
...
  /sales:
    get:
      consumes:
      - application/json
      description: Retrieve a list of sales with optional filters
      parameters:
      - in: query
        name: client_id
        type: string
      - in: query
        name: end_date
        type: string
      - in: query
        name: sold_by
        type: string
      - in: query
        name: start_date
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.SaleList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: List all sales
      tags:
      - Sales
    post:
      consumes:
      - application/json
      description: Create a new sale with the provided details
      parameters:
      - description: Sale data
        in: body
        name: Sale
        required: true
        schema:
          $ref: '#/definitions/entity.Sale'
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/products.SaleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a new sale
      tags:
      - Sales
  /sales/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a sale by ID
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete a sale
      tags:
      - Sales
    get:
      consumes:
      - application/json
      description: Retrieve a sale by ID
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.SaleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Get a sale
      tags:
      - Sales
    put:
      consumes:
      - application/json
      description: Update the details of an existing sale by ID
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated sale data
        in: body
        name: Sale
        required: true
        schema:
          $ref: '#/definitions/entity.SaleUpdate'
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.SaleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Update an existing sale
      tags:
      - Sales
  /sales/calculate:
    post:
      consumes:
      - application/json
      description: Calculate the total sales based on the sale request
      parameters:
      - description: Sale data
        in: body
        name: Sale
        required: true
        schema:
          $ref: '#/definitions/entity.Sale'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.SaleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Calculate total sales
      tags:
      - Sales
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"